Solving circular import and refactoring
