+++ src/tsdf/__init__.py
@@ -3,14 +3,23 @@ from importlib.metadata import version
 
 __version__ = version("tsdf")
 
-from .io import (
+from .load_tsdf import (
     load_metadata_file,
     load_metadata_from_path,
     load_metadata_string,
+    load_metadata_legacy_file
+)
+from .write_tsdf import (
+    write_metadata,
+)
+    
+from .io_binary import (
     load_binary_from_metadata,
     load_binary_file,
     get_metadata_from_ndarray,
     write_binary_file,
-    write_metadata,
 )
+
+
+    
 from .tsdf_metadata import TSDFMetadata
